requirements.txt
APScheduler==3.6.3
beautifulsoup4==4.8.2
certifi==2019.11.28
chardet==3.0.4
click==7.1.1
fake-useragent==0.1.11
feedgen==0.9.0
Flask==1.1.1
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.1
lxml==4.5.0
MarkupSafe==1.1.1
python-dateutil==2.8.1
pytz==2019.3
requests==2.23.0
six==1.14.0
soupsieve==2.0
tzlocal==2.0.0
urllib3==1.25.8
Werkzeug==1.0.0
